fix fp254impl issues
