+++ bitvm/src/bn254/fp254impl.rs
@@ -1,14 +1,11 @@
 use crate::bigint::add::limb_add_carry;
 use crate::bigint::sub::limb_sub_borrow;
 use crate::bigint::U254;
-use crate::bn254::fq::Fq;
-use crate::bn254::utils::Hint;
 use crate::treepp::*;
 use ark_ff::PrimeField;
 use bitcoin_script::script;
-use num_bigint::{BigInt, BigUint};
+use num_bigint::BigUint;
 use num_traits::Num;
-use std::str::FromStr;
 use std::sync::OnceLock;
 
 #[allow(clippy::declare_interior_mutable_const)]
@@ -205,9 +202,7 @@ pub trait Fp254Impl {
                     // ⋯ A₈ B₈ A₇ B₇ A₆ B₆ A₅ B₅ A₄ B₄ A₃ B₃ A₂ B₂ C₁⁺ (B₁+C₀)+A₁ 2²⁹ C₀⁻ M₁
                     OP_ADD
                     // ⋯ A₈ B₈ A₇ B₇ A₆ B₆ A₅ B₅ A₄ B₄ A₃ B₃ A₂ B₂ C₁⁺ (B₁+C₀)+A₁ 2²⁹ C₀⁻+M₁
-                    OP_ROT
                     OP_SWAP
-                    OP_ROT
                     // ⋯ A₈ B₈ A₇ B₇ A₆ B₆ A₅ B₅ A₄ B₄ A₃ B₃ A₂ B₂ C₁⁺ (B₁+C₀)+A₁ C₀⁻+M₁ 2²⁹
                     limb_sub_borrow
                     OP_TOALTSTACK
@@ -275,51 +270,25 @@ pub trait Fp254Impl {
             { Self::roll(a) }
             { Self::is_zero_keep_element(0) }
             OP_NOTIF
-                // ⋯ A₈ A₇ A₆ A₅ A₄ A₃ A₂ A₁ A₀
-                { Self::MODULUS_LIMBS[0] } OP_SWAP { 0x20000000 }
-                limb_sub_borrow OP_TOALTSTACK
-                // ⋯ A₈ A₇ A₆ A₅ A₄ A₃ A₂ A₁ 2²⁹ C₀⁻ | M₀-A₀ ⋯
-                OP_ROT OP_ADD
-                // ⋯ A₈ A₇ A₆ A₅ A₄ A₃ A₂ 2²⁹ C₀⁻+A₁
-                { Self::MODULUS_LIMBS[1] } OP_SWAP OP_ROT
-                limb_sub_borrow OP_TOALTSTACK
-                // ⋯ A₈ A₇ A₆ A₅ A₄ A₃ A₂ 2²⁹ C₁⁻ | M₁-(C₀⁻+A₁) ⋯
-                OP_ROT OP_ADD
-                // ⋯ A₈ A₇ A₆ A₅ A₄ A₃ 2²⁹ C₁⁻+A₂
-                { Self::MODULUS_LIMBS[2] } OP_SWAP OP_ROT
-                limb_sub_borrow OP_TOALTSTACK
-                // ⋯ A₈ A₇ A₆ A₅ A₄ A₃ 2²⁹ C₂⁻ | M₂-(C₁⁻+A₂) ⋯
-                OP_ROT OP_ADD
-                // ⋯ A₈ A₇ A₆ A₅ A₄ 2²⁹ C₂⁻+A₃
-                { Self::MODULUS_LIMBS[3] } OP_SWAP OP_ROT
-                limb_sub_borrow OP_TOALTSTACK
-                // ⋯ A₈ A₇ A₆ A₅ A₄ 2²⁹ C₃⁻ | M₃-(C₂⁻+A₃) ⋯
-                OP_ROT OP_ADD
-                // ⋯ A₈ A₇ A₆ A₅ 2²⁹ C₃⁻+A₄
-                { Self::MODULUS_LIMBS[4] } OP_SWAP OP_ROT
-                limb_sub_borrow OP_TOALTSTACK
-                // ⋯ A₈ A₇ A₆ A₅ 2²⁹ C₄⁻ | M₄-(C₃⁻+A₄) ⋯
-                OP_ROT OP_ADD
-                // ⋯ A₈ A₇ A₆ 2²⁹ C₄⁻+A₅
-                { Self::MODULUS_LIMBS[5] } OP_SWAP OP_ROT
-                limb_sub_borrow OP_TOALTSTACK
-                // ⋯ A₈ A₇ A₆ 2²⁹ C₅⁻ | M₅-(C₄⁻+A₅) ⋯
-                OP_ROT OP_ADD
-                // ⋯ A₈ A₇ 2²⁹ C₅⁻+A₆
-                { Self::MODULUS_LIMBS[6] } OP_SWAP OP_ROT
-                limb_sub_borrow OP_TOALTSTACK
-                // ⋯ A₈ A₇ 2²⁹ C₆⁻ | M₆-(C₅⁻+A₆) ⋯
-                OP_ROT OP_ADD
-                // ⋯ A₈ 2²⁹ C₆⁻+A₇
-                { Self::MODULUS_LIMBS[7] } OP_SWAP OP_ROT
-                limb_sub_borrow OP_TOALTSTACK
-                // ⋯ A₈ 2²⁹ C₇⁻ | M₇-(C₆⁻+A₇) ⋯
-                OP_NIP OP_ADD
-                // ⋯ C₇⁻+A₈
-                { Self::MODULUS_LIMBS[8] } OP_SWAP OP_SUB
-                // ⋯ M₈-(C₇⁻+A₈)
-                OP_FROMALTSTACK OP_FROMALTSTACK OP_FROMALTSTACK OP_FROMALTSTACK
-                OP_FROMALTSTACK OP_FROMALTSTACK OP_FROMALTSTACK OP_FROMALTSTACK
+                for i in 0..Self::N_LIMBS-1 {
+                    { Self::MODULUS_LIMBS[i as usize] } OP_SWAP
+                    if i == 0 {
+                        { 2_usize.pow(U254::LIMB_SIZE) }
+                    } else {
+                        OP_ROT
+                    }
+                    limb_sub_borrow OP_TOALTSTACK
+                    if i == Self::N_LIMBS-2 {
+                        OP_NIP
+                    } else {
+                        OP_ROT
+                    }
+                    OP_ADD
+                }
+                { Self::MODULUS_LIMBS[Self::N_LIMBS as usize - 1] } OP_SWAP OP_SUB
+                for _ in 0..Self::N_LIMBS-1 {
+                    OP_FROMALTSTACK
+                }
             OP_ENDIF
         }
     }
@@ -558,371 +527,5 @@ pub trait Fp254Impl {
                 OP_DROP
             OP_ENDIF
         }
-    }
-
-    fn hinted_mul(
-        mut a_depth: u32,
-        mut a: ark_bn254::Fq,
-        mut b_depth: u32,
-        mut b: ark_bn254::Fq,
-    ) -> (Script, Vec<Hint>) {
-        assert_ne!(a_depth, b_depth);
-        if a_depth > b_depth {
-            (a_depth, b_depth) = (b_depth, a_depth);
-            (a, b) = (b, a);
-        }
-
-        let mut hints = Vec::new();
-        let x = BigInt::from_str(&a.to_string()).unwrap();
-        let y = BigInt::from_str(&b.to_string()).unwrap();
-        let modulus = &Fq::modulus_as_bigint();
-        let q = (x * y) / modulus;
-
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::roll(a_depth + 1) }
-            { Fq::roll(b_depth + 1) }
-            { Fq::tmul() }
-        };
-        hints.push(Hint::BigIntegerTmulLC1(q));
-
-        (script, hints)
-    }
-
-    // TODO: Optimize by using the constant feature
-    fn hinted_mul_by_constant(a: ark_bn254::Fq, constant: &ark_bn254::Fq) -> (Script, Vec<Hint>) {
-        let mut hints = Vec::new();
-        let x = BigInt::from_str(&a.to_string()).unwrap();
-        let y = BigInt::from_str(&constant.to_string()).unwrap();
-        let modulus = &Fq::modulus_as_bigint();
-        let q = (x * y) / modulus;
-
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::roll(1) }
-            { Fq::push(*constant) }
-            { Fq::tmul() }
-        };
-        hints.push(Hint::BigIntegerTmulLC1(q));
-
-        (script, hints)
-    }
-
-    fn hinted_mul_keep_element(
-        mut a_depth: u32,
-        mut a: ark_bn254::Fq,
-        mut b_depth: u32,
-        mut b: ark_bn254::Fq,
-    ) -> (Script, Vec<Hint>) {
-        assert_ne!(a_depth, b_depth);
-        if a_depth > b_depth {
-            (a_depth, b_depth) = (b_depth, a_depth);
-            (a, b) = (b, a);
-        }
-
-        let mut hints = Vec::new();
-        let x = BigInt::from_str(&a.to_string()).unwrap();
-        let y = BigInt::from_str(&b.to_string()).unwrap();
-        let modulus = &Fq::modulus_as_bigint();
-        let q = (x * y) / modulus;
-
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::copy(a_depth + 1) }
-            { Fq::copy(b_depth + 2) }
-            { Fq::tmul() }
-        };
-        hints.push(Hint::BigIntegerTmulLC1(q));
-
-        (script, hints)
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn hinted_mul_lc2(
-        a_depth: u32,
-        a: ark_bn254::Fq,
-        b_depth: u32,
-        b: ark_bn254::Fq,
-        c_depth: u32,
-        c: ark_bn254::Fq,
-        d_depth: u32,
-        d: ark_bn254::Fq,
-    ) -> (Script, Vec<Hint>) {
-        assert!(a_depth > b_depth && b_depth > c_depth && c_depth > d_depth);
-
-        let mut hints = Vec::new();
-
-        let modulus = &Fq::modulus_as_bigint();
-
-        let x = BigInt::from_str(&a.to_string()).unwrap();
-        let y = BigInt::from_str(&b.to_string()).unwrap();
-        let z = BigInt::from_str(&c.to_string()).unwrap();
-        let w = BigInt::from_str(&d.to_string()).unwrap();
-
-        let q = (x * z + y * w) / modulus;
-
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::roll(a_depth + 1) }
-            { Fq::roll(b_depth + 2) }
-            { Fq::roll(c_depth + 3) }
-            { Fq::roll(d_depth + 4) }
-            { Fq::tmul_lc2() }
-        };
-        hints.push(Hint::BigIntegerTmulLC2(q));
-
-        (script, hints)
-    }
-
-    // Assumes tmul hint (1 BigInteger) at the top of stack
-    // and operands (a, b, c, d) at stack depths (a_depth, b_depth, c_depth, d_depth)
-    // Computes r = a * c + b * d (mod p)
-    #[allow(clippy::too_many_arguments)]
-    fn hinted_mul_lc2_w4(
-        a_depth: u32,
-        a: ark_bn254::Fq,
-        b_depth: u32,
-        b: ark_bn254::Fq,
-        c_depth: u32,
-        c: ark_bn254::Fq,
-        d_depth: u32,
-        d: ark_bn254::Fq,
-    ) -> (Script, Vec<Hint>) {
-        assert!(a_depth > b_depth && b_depth > c_depth && c_depth > d_depth);
-
-        let mut hints = Vec::with_capacity(1);
-
-        let modulus = &Fq::modulus_as_bigint();
-
-        let x = BigInt::from_str(&a.to_string()).unwrap();
-        let y = BigInt::from_str(&b.to_string()).unwrap();
-        let z = BigInt::from_str(&c.to_string()).unwrap();
-        let w = BigInt::from_str(&d.to_string()).unwrap();
-
-        let q = (x * z + y * w) / modulus;
-
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::roll(a_depth + 1) }
-            { Fq::roll(b_depth + 2) }
-            { Fq::roll(c_depth + 3) }
-            { Fq::roll(d_depth + 4) }
-            { Fq::tmul_lc2_w4() }
-        };
-        hints.push(Hint::BigIntegerTmulLC2(q));
-
-        (script, hints)
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn hinted_mul_lc4(
-        a_depth: u32,
-        a: ark_bn254::Fq,
-        b_depth: u32,
-        b: ark_bn254::Fq,
-        c_depth: u32,
-        c: ark_bn254::Fq,
-        d_depth: u32,
-        d: ark_bn254::Fq,
-
-        e_depth: u32,
-        e: ark_bn254::Fq,
-        f_depth: u32,
-        f: ark_bn254::Fq,
-        g_depth: u32,
-        g: ark_bn254::Fq,
-        h_depth: u32,
-        h: ark_bn254::Fq,
-    ) -> (Script, Vec<Hint>) {
-        assert!(
-            a_depth > b_depth
-                && b_depth > c_depth
-                && c_depth > d_depth
-                && d_depth > e_depth
-                && e_depth > f_depth
-                && f_depth > g_depth
-                && g_depth > h_depth
-        );
-
-        let mut hints = Vec::new();
-
-        let modulus = &Fq::modulus_as_bigint();
-
-        let x1 = BigInt::from_str(&a.to_string()).unwrap();
-        let y1 = BigInt::from_str(&b.to_string()).unwrap();
-        let z1 = BigInt::from_str(&c.to_string()).unwrap();
-        let w1 = BigInt::from_str(&d.to_string()).unwrap();
-
-        let x2 = BigInt::from_str(&e.to_string()).unwrap();
-        let y2 = BigInt::from_str(&f.to_string()).unwrap();
-        let z2 = BigInt::from_str(&g.to_string()).unwrap();
-        let w2 = BigInt::from_str(&h.to_string()).unwrap();
-
-        let q = (x1 * x2 + y1 * y2 + z1 * z2 + w1 * w2) / modulus;
-
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { fq_push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::roll(a_depth + 1) }
-            { Fq::roll(b_depth + 2) }
-            { Fq::roll(c_depth + 3) }
-            { Fq::roll(d_depth + 4) }
-            { Fq::roll(e_depth + 5) }
-            { Fq::roll(f_depth + 6) }
-            { Fq::roll(g_depth + 7) }
-            { Fq::roll(h_depth + 8) }
-            { Fq::tmul_lc4() }
-        };
-        hints.push(Hint::BigIntegerTmulLC4(q));
-
-        (script, hints)
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn hinted_mul_lc2_keep_elements(
-        a_depth: u32,
-        a: ark_bn254::Fq,
-        b_depth: u32,
-        b: ark_bn254::Fq,
-        c_depth: u32,
-        c: ark_bn254::Fq,
-        d_depth: u32,
-        d: ark_bn254::Fq,
-    ) -> (Script, Vec<Hint>) {
-        assert!(a_depth > b_depth && b_depth > c_depth && c_depth > d_depth);
-
-        let mut hints = Vec::new();
-
-        let modulus = &Fq::modulus_as_bigint();
-
-        let x = BigInt::from_str(&a.to_string()).unwrap();
-        let y = BigInt::from_str(&b.to_string()).unwrap();
-        let z = BigInt::from_str(&c.to_string()).unwrap();
-        let w = BigInt::from_str(&d.to_string()).unwrap();
-
-        let q = (x * z + y * w) / modulus;
-
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::copy(a_depth + 1) }
-            { Fq::copy(b_depth + 2) }
-            { Fq::copy(c_depth + 3) }
-            { Fq::copy(d_depth + 4) }
-            { Fq::tmul_lc2() }
-        };
-        hints.push(Hint::BigIntegerTmulLC2(q));
-
-        (script, hints)
-    }
-
-    // Same as hinted_mul_lc2_keep_elements(), except retains operands (a, b, c, d) on stack
-    #[allow(clippy::too_many_arguments)]
-    fn hinted_mul_lc2_keep_elements_w4(
-        a_depth: u32,
-        a: ark_bn254::Fq,
-        b_depth: u32,
-        b: ark_bn254::Fq,
-        c_depth: u32,
-        c: ark_bn254::Fq,
-        d_depth: u32,
-        d: ark_bn254::Fq,
-    ) -> (Script, Vec<Hint>) {
-        assert!(a_depth > b_depth && b_depth > c_depth && c_depth > d_depth);
-
-        let mut hints = Vec::with_capacity(1);
-
-        let modulus = &Fq::modulus_as_bigint();
-
-        let x = BigInt::from_str(&a.to_string()).unwrap();
-        let y = BigInt::from_str(&b.to_string()).unwrap();
-        let z = BigInt::from_str(&c.to_string()).unwrap();
-        let w = BigInt::from_str(&d.to_string()).unwrap();
-
-        let q = (x * z + y * w) / modulus;
-
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::copy(a_depth + 1) }
-            { Fq::copy(b_depth + 2) }
-            { Fq::copy(c_depth + 3) }
-            { Fq::copy(d_depth + 4) }
-            { Fq::tmul_lc2_w4() }
-        };
-        hints.push(Hint::BigIntegerTmulLC2(q));
-
-        (script, hints)
-    }
-
-    // TODO: Optimize using the sqaure feature
-    fn hinted_square(a: ark_bn254::Fq) -> (Script, Vec<Hint>) {
-        let mut hints = Vec::new();
-        let x = &BigInt::from_str(&a.to_string()).unwrap();
-        let modulus = &Fq::modulus_as_bigint();
-        let q = (x * x) / modulus;
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            { Fq::roll(1) }
-            { Fq::copy(0) }
-            { Fq::tmul() }
-        };
-        hints.push(Hint::BigIntegerTmulLC1(q));
-
-        (script, hints)
-    }
-
-    fn hinted_inv(a: ark_bn254::Fq) -> (Script, Vec<Hint>) {
-        let mut hints = Vec::new();
-        let x = &BigInt::from_str(&a.to_string()).unwrap();
-        let modulus = &Fq::modulus_as_bigint();
-        let y = &x.modinv(modulus).unwrap();
-        let q = (x * y) / modulus;
-        let script = script! {
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            for _ in 0..Self::N_LIMBS {
-                OP_DEPTH OP_1SUB OP_ROLL // hints
-            }
-            // { Fq::push(ark_bn254::Fq::from_str(&y.to_string()).unwrap()) }
-            // { Fq::push(ark_bn254::Fq::from_str(&q.to_string()).unwrap()) }
-            // x, y, q
-            { Fq::roll(2) }
-            { Fq::copy(2) }
-            // y, q, x, y
-            { Fq::tmul() }
-            // y, 1
-            { Fq::push_one() }
-            { Fq::equalverify(1, 0) }
-        };
-        hints.push(Hint::Fq(ark_bn254::Fq::from_str(&y.to_string()).unwrap()));
-        hints.push(Hint::BigIntegerTmulLC1(q));
-
-        (script, hints)
     }
 }
